requirements.txt
absl-py==0.7.1
astor==0.8.0
Click==7.0
colorcet==2.0.1
Flask==1.0.3
Flask-Cors==3.0.8
gast==0.2.2
gevent==1.4.0
greenlet==0.4.15
grpcio==1.21.1
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
nltk==3.4.3
numpy==1.16.4
param==1.9.1
Pillow==6.0.0
protobuf==3.8.0
pyct==0.4.6
runway-python==0.1.0
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
tensorflow-hub==0.4.0
termcolor==1.1.0
tf-sentencepiece==0.1.82
Werkzeug==0.15.4
wget==3.2
